Refactory some components
