+++ src/components/selectDividerCo.js
@@ -1,7 +1,7 @@
 import React from "react"
 import Divider from "@material-ui/core/Divider"
 import { makeStyles } from "@material-ui/core/styles"
-import theme from "../../theme"
+import theme from "../theme"
 // We can inject some CSS into the DOM.
 const useStyles = makeStyles({
   root: {
