+++ src/section/SelectSection.js
@@ -3,8 +3,8 @@ import Grid from "@material-ui/core/Grid"
 import Box from "@material-ui/core/Box"
 import LanguageIcon from "@material-ui/icons/Language"
 import TypographyCo from "../components/typographyCo"
-import MenuItemCo from "../components/secondaryComponents/menuItemCo"
-import SelectDividerCo from "../components/secondaryComponents/selectDividerCo"
+import MenuItemCo from "../components/menuItemCo"
+import SelectDividerCo from "../components/selectDividerCo"
 import SelectCo from "../components/selectCo"
 import states from "../assets/states"
 import theme from "../theme"
